feat: installing tailwind, moving node-spec to different file
