+++ docusaurus.config.js
@@ -45,6 +45,17 @@ const config = {
     ],
   ],
   plugins: [
+    async function tailwind(context, options) {
+      return {
+        name: "docusaurus-tailwindcss",
+        configurePostCss(postcssOptions) {
+          // Appends TailwindCSS and AutoPrefixer.
+          postcssOptions.plugins.push(require("tailwindcss"));
+          postcssOptions.plugins.push(require("autoprefixer"));
+          return postcssOptions;
+        },
+      };
+    },
     ["drawio", {}],
     ['@docusaurus/plugin-client-redirects', {
       redirects: [
