+++ package.json
@@ -22,14 +22,19 @@
     "@docusaurus/core": "^2.2.0",
     "@docusaurus/plugin-client-redirects": "^2.2.0",
     "@docusaurus/preset-classic": "^2.2.0",
+    "@heroicons/react": "^2.0.15",
     "@mdx-js/react": "^1.6.22",
+    "autoprefixer": "^10.4.13",
     "clsx": "^1.1.1",
     "docusaurus-plugin-drawio": "^0.1.7",
     "dotenv": "^16.0.3",
+    "postcss": "^8.4.21",
     "prism-react-renderer": "^1.2.1",
     "raw-loader": "^4.0.2",
     "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "react-dom": "^17.0.1",
+    "react-syntax-highlighter": "^15.5.0",
+    "tailwindcss": "^3.2.6"
   },
   "browserslist": {
     "production": [
